requirements.txt
Flask==3.0.0
yfinance==0.2.32
numpy==1.26.4
scikit-learn==1.4.1
boto3==1.34.0
python-dotenv==1.0.0
